feat: add plugin hooks for email details, headers, and source
